Move format, create_pool, and create_buffer to wl_shm.rs
